src/requirements.txt
taipy==3.0.0.dev3
opencv-python-headless==4.7.0.72
opencv-contrib-python-headless==4.7.0.72
pillow